Calls Route handlers from main thread.
Previously Perfidy was calling any Route handlers from the thread of the NIO event loop that triggered them. But these handlers are, more often than not, closures created in tests that interacted with UI elements. So it’s exceedingly rare to see one that doesn’t immediately jump to `DispatchQueue.main` — or *should* and forgets to do so. So Perfidy does that for us now.