Removes dependecy on Medea
Some of Medea is now done better with codable. Some of it is unnecessary becasue of improvements to Swift. Of the rest (mostly around validation), there's just not enough to justify a dependency.

Though everything should still work, this is technically a breaking change as `JSONObject` and `JSONArray` types and various errors are now in the `Perfidy` namespace.