Makes `start` implicit with initialization.
As commonly used, it was rare to create a `FakeServer` and not immediately `start` it. So this step is now a part of initialization.

It would make sense, then, to `stop` in `deinit`, right? Except that swift tends to favor optimization over predictable object lifespans. Without `withExtendedLifetime` there’s always the chance of the server getting deinit early, and we’ve already seen issues with the `runWith` closure interacting confusingly with asynchronous tests resulting in early termination of the server. Best for everyone to keep `stop` explicit.